+++ backend/src/main/java/com/digginroom/digginroom/controller/GlobalExceptionHandler.java
@@ -4,7 +4,7 @@ import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
-import com.digginroom.digginroom.controller.dto.ErrorResponse;
+import com.digginroom.digginroom.service.dto.ErrorResponse;
 import com.digginroom.digginroom.exception.DigginRoomException;
 import java.time.LocalDateTime;
 import java.util.stream.Collectors;
