+++ backend/src/main/java/com/digginroom/digginroom/controller/MemberLoginController.java
@@ -1,8 +1,8 @@
 package com.digginroom.digginroom.controller;
 
-import com.digginroom.digginroom.controller.dto.IdTokenRequest;
-import com.digginroom.digginroom.controller.dto.MemberLoginRequest;
-import com.digginroom.digginroom.controller.dto.MemberLoginResponse;
+import com.digginroom.digginroom.service.dto.IdTokenRequest;
+import com.digginroom.digginroom.service.dto.MemberLoginRequest;
+import com.digginroom.digginroom.service.dto.MemberLoginResponse;
 import com.digginroom.digginroom.service.MemberService;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
