+++ backend/src/main/java/com/digginroom/digginroom/controller/RoomController.java
@@ -1,8 +1,8 @@
 package com.digginroom.digginroom.controller;
 
-import com.digginroom.digginroom.controller.dto.RoomRequest;
-import com.digginroom.digginroom.controller.dto.RoomResponse;
-import com.digginroom.digginroom.controller.dto.RoomsResponse;
+import com.digginroom.digginroom.service.dto.RoomRequest;
+import com.digginroom.digginroom.service.dto.RoomResponse;
+import com.digginroom.digginroom.service.dto.RoomsResponse;
 import com.digginroom.digginroom.service.RoomService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
