--- backend/src/main/java/com/digginroom/digginroom/controller/dto/RoomRequest.java
@@ -1,4 +1,0 @@
-package com.digginroom.digginroom.controller.dto;
-
-public record RoomRequest(Long roomId) {
-}
