+++ backend/src/main/java/com/digginroom/digginroom/service/dto/CommentResponse.java
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.controller.dto;
+package com.digginroom.digginroom.service.dto;
 
 import com.digginroom.digginroom.domain.comment.Comment;
 import java.time.LocalDateTime;
