+++ backend/src/main/java/com/digginroom/digginroom/service/dto/CommentsResponse.java
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.controller.dto;
+package com.digginroom.digginroom.service.dto;
 
 import java.util.List;
