+++ backend/src/main/java/com/digginroom/digginroom/service/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.controller.dto;
+package com.digginroom.digginroom.service.dto;
 
 import java.time.LocalDateTime;
