+++ backend/src/main/java/com/digginroom/digginroom/service/dto/MemberSaveRequest.java
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.controller.dto;
+package com.digginroom.digginroom.service.dto;
 
 import com.digginroom.digginroom.domain.member.Member;
 import jakarta.validation.constraints.NotBlank;
