+++ backend/src/main/java/com/digginroom/digginroom/service/dto/RoomRequest.java
@@ -1,0 +1,4 @@
+package com.digginroom.digginroom.service.dto;
+
+public record RoomRequest(Long roomId) {
+}
