+++ backend/src/main/java/com/digginroom/digginroom/service/dto/RoomResponse.java
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.controller.dto;
+package com.digginroom.digginroom.service.dto;
 
 public record RoomResponse(Long roomId, String videoId, boolean isScrapped, Long scrapCount, TrackResponse track) {
 }
